+++ src/components/admin/overNight/ReadAbsence.tsx
@@ -3,6 +3,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import { privateApi } from "../../../services/customAxios";
 import { ListBtn } from "../../table/Table";
 import { AbsenceType } from "../../../types/admin";
+import { useMutation, useQuery } from "@tanstack/react-query";
 
 function ReadAbsence() {
   // 글 ID 값
@@ -12,29 +13,38 @@ function ReadAbsence() {
   const [absence, setAbsence] = useState<AbsenceType>();
   const navigate = useNavigate();
 
-  // 페이지 렌더링 시
+  // 외출, 외박 데이터 get Api
+  const getAbsenceDataApi = async () => {
+    const response = await privateApi.get(`/api/absence/${id}`);
+
+    return response.data;
+  };
+
+  // 외출, 외박 거절 Api
+  const patchAbsenceApi = async (id: string) => {
+    const response = await privateApi.patch(`/api/absence/reject/${id}`);
+
+    return response.data;
+  };
+
+  // 외출, 외박 query
+  const { data } = useQuery({
+    queryKey: ["absenceData"],
+    queryFn: getAbsenceDataApi,
+  });
+
   useEffect(() => {
-    getAbsenceData();
-  }, []);
+    if (data) setAbsence(data.stay_out);
+  }, [data]);
 
-  // 외박,외출 가져오기
-  const getAbsenceData = async () => {
-    try {
-      const getNotice = await privateApi.get(`/api/absence/${id}`);
-      setAbsence(getNotice.data.stay_out);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // 외출,외박 거절하기
-  const patchAbsenceData = async (id: string) => {
-    try {
-      await privateApi.patch(`/api/absence/reject/${id}`);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  // 외출, 외박 거절 mutation
+  const { mutate: patchAbsenceMutation } = useMutation({
+    mutationFn: (id: string) => patchAbsenceApi(id),
+    // Api 연결 성공
+    onSuccess() {
+      navigate("/main/admin/stayOut");
+    },
+  });
 
   return (
     <div className="flex flex-col gap-4">
@@ -67,9 +77,7 @@ function ReadAbsence() {
             if (id) {
               if (window.confirm("거절하시겠습니까?")) {
                 alert("거절되었습니다");
-                patchAbsenceData(id).then(() => {
-                  navigate("/main/admin/stayOut");
-                });
+                patchAbsenceMutation(id);
               } else {
                 alert("취소되었습니다.");
               }
