--- src/pages/admin/Admin.tsx
@@ -1,12 +1,0 @@
-import MainContainer from "../../components/MainContainer";
-import AdminMain from "../../components/admin/main/AdminMain";
-
-function Admin() {
-  return (
-    <MainContainer>
-      <AdminMain />
-    </MainContainer>
-  );
-}
-
-export default Admin;
