+++ src/routes/AdminRoutes.tsx
@@ -1,5 +1,5 @@
 import { Route, Routes } from "react-router-dom";
-import Admin from "../pages/admin/Admin";
+import Notice from "../pages/admin/Notice";
 import Repair from "../pages/admin/Repair";
 import StayOut from "../pages/admin/StayOut";
 import MeetingRoom from "../pages/admin/MeetingRoom";
@@ -11,7 +11,7 @@ import Modifying from "../pages/post/Modifying";
 function AdminRoutes() {
   return (
     <Routes>
-      <Route path="/notice" element={<Admin />} />
+      <Route path="/notice" element={<Notice />} />
       <Route path="/repair" element={<Repair />} />
       <Route path="/stayOut" element={<StayOut />} />
       <Route path="/meetingRoom" element={<MeetingRoom />} />
